+++ build.properties
@@ -1,5 +1,5 @@
 # Project specific information
-package=org.usfirst.frc.team4571.robot
+package=org.usfirst.frc.team7788.robot
 robot.class=${package}.Robot
 simulation.world.file=/usr/share/frcsim/worlds/GearsBotDemo.world
 #Uncomment and point at user libraries to include them in the build. Do not put libraries in the \wpilib\java folder, this folder is completely overwritten on plugin update.
