--- src/org/usfirst/frc/team4571/robot/subsystems/PIDsource.java
@@ -1,5 +1,0 @@
-package org.usfirst.frc.team4571.robot.subsystems;
-
-public interface PIDsource {
-
-}
