+++ src/org/usfirst/frc/team7788/robot/OI.java
@@ -1,8 +1,8 @@
-package org.usfirst.frc.team4571.robot;
+package org.usfirst.frc.team7788.robot;
+
+import org.usfirst.frc.team7788.robot.commands.ExampleCommand;
 
 import edu.wpi.first.wpilibj.buttons.Button;
-
-import org.usfirst.frc.team4571.robot.commands.ExampleCommand;
 
 /**
  * This class is the glue that binds the controls on the physical operator
