+++ src/org/usfirst/frc/team7788/robot/RobotConstants.java
@@ -1,4 +1,4 @@
-package org.usfirst.frc.team4571.robot;
+package org.usfirst.frc.team7788.robot;
 
 
 public class RobotConstants {
