+++ src/org/usfirst/frc/team7788/robot/RobotJoystick.java
@@ -1,4 +1,4 @@
-package org.usfirst.frc.team4571.robot;
+package org.usfirst.frc.team7788.robot;
 
 import edu.wpi.first.wpilibj.Joystick;
 import edu.wpi.first.wpilibj.buttons.Button;
