+++ src/org/usfirst/frc/team7788/robot/commands/AutonomousDriveCommand2.java
@@ -1,14 +1,15 @@
-package org.usfirst.frc.team4571.robot.commands;
+package org.usfirst.frc.team7788.robot.commands;
 
-import org.usfirst.frc.team4571.robot.Robot;
+import org.usfirst.frc.team7788.robot.Robot;
+
 import edu.wpi.first.wpilibj.command.Command;
 
 /**
  *
  */
-public class AutonomousDriveCommand extends Command {
+public class AutonomousDriveCommand2 extends Command {
 	public boolean driveStop = false;
-	public AutonomousDriveCommand() {
+	public AutonomousDriveCommand2() {
 		requires(Robot.TANK_DRIVE_SUBSYSTEM);
 		// Use requires() here to declare subsystem dependencies
 		// eg. requires(chassis);
@@ -20,19 +21,9 @@ public class AutonomousDriveCommand extends Command {
 	// Called repeatedly when this Command is scheduled to run
 	protected void execute() {
 		long startFowardTime = System.currentTimeMillis();
-		long finishFowardTime = 10000;
+		long finishFowardTime = 1800000;
 		while( System.currentTimeMillis() - startFowardTime <= finishFowardTime ){
-			Robot.TANK_DRIVE_SUBSYSTEM.drive(1,1,true);
-		}
-		long fowardTimeStop = System.currentTimeMillis();
-		long finishTimeStop = 2000;
-		while(System.currentTimeMillis() - fowardTimeStop <= finishTimeStop ){
-			Robot.TANK_DRIVE_SUBSYSTEM.stop();
-		}
-		long startReverseTime = System.currentTimeMillis();
-		long finishReverseTime = 5000;
-		while( System.currentTimeMillis() - startReverseTime <= finishReverseTime ){
-			Robot.TANK_DRIVE_SUBSYSTEM.drive(-1,-1,true);
+			Robot.TANK_DRIVE_SUBSYSTEM.drive(.3,.3,true);
 			driveStop = true;
 		}
