+++ src/org/usfirst/frc/team7788/robot/commands/ExampleCommand.java
@@ -1,4 +1,4 @@
-package org.usfirst.frc.team4571.robot.commands;
+package org.usfirst.frc.team7788.robot.commands;
 
 import edu.wpi.first.wpilibj.command.Command;
