+++ src/org/usfirst/frc/team7788/robot/commands/Gearpneumaticscommand.java
@@ -1,6 +1,6 @@
-package org.usfirst.frc.team4571.robot.commands;
+package org.usfirst.frc.team7788.robot.commands;
 
-import org.usfirst.frc.team4571.robot.Robot;
+import org.usfirst.frc.team7788.robot.Robot;
 
 import edu.wpi.first.wpilibj.command.Command;
 
@@ -8,6 +8,7 @@ public class Gearpneumaticscommand extends Command {
 
 	
 	private boolean IsGearPushedOut;
+	
 	public  Gearpneumaticscommand (){
 		requires(Robot.GEAR_SUBSYSTEM);
 	}
