+++ src/org/usfirst/frc/team7788/robot/subsystems/PID/DistanceOutput.java
@@ -1,4 +1,4 @@
-package org.usfirst.frc.team4571.robot.subsystems.PID;
+package org.usfirst.frc.team7788.robot.subsystems.PID;
 
 import edu.wpi.first.wpilibj.PIDOutput;
