+++ src/org/usfirst/frc/team7788/robot/subsystems/PIDsource.java
@@ -1,0 +1,5 @@
+package org.usfirst.frc.team7788.robot.subsystems;
+
+public interface PIDsource {
+
+}
