+++ src/org/usfirst/frc/team7788/robot/subsystems/ShooterOutput.java
@@ -1,4 +1,4 @@
-package org.usfirst.frc.team4571.robot.subsystems;
+package org.usfirst.frc.team7788.robot.subsystems;
 
 import com.ctre.CANTalon;
