+++ src/org/usfirst/frc/team7788/robot/subsystems/ShooterSource.java
@@ -1,4 +1,4 @@
-package org.usfirst.frc.team4571.robot.subsystems;
+package org.usfirst.frc.team7788.robot.subsystems;
 import edu.wpi.first.wpilibj.Encoder;
 import edu.wpi.first.wpilibj.PIDSource;
 import edu.wpi.first.wpilibj.PIDSourceType;
