+++ src/org/usfirst/frc/team7788/robot/subsystems/ShooterSubsystem.java
@@ -1,8 +1,8 @@
-package org.usfirst.frc.team4571.robot.subsystems;
+package org.usfirst.frc.team7788.robot.subsystems;
 
-import org.usfirst.frc.team4571.robot.RobotConstants;
-import org.usfirst.frc.team4571.robot.subsystems.ShooterSource;
-import org.usfirst.frc.team4571.robot.subsystems.ShooterOutput;
+import org.usfirst.frc.team7788.robot.RobotConstants;
+import org.usfirst.frc.team7788.robot.subsystems.ShooterOutput;
+import org.usfirst.frc.team7788.robot.subsystems.ShooterSource;
 
 import com.ctre.CANTalon;
