+++ src/org/usfirst/frc/team7788/robot/subsystems/averagedistanceEncoder.java
@@ -1,4 +1,4 @@
-package org.usfirst.frc.team4571.robot.subsystems;
+package org.usfirst.frc.team7788.robot.subsystems;
 
 import edu.wpi.first.wpilibj.PIDSource;
